requirements.txt
discord.py==2.3.2
Flask==2.3.3
python-dotenv==1.0.1
openai>=1.3.7
httpx==0.27.0
